refactor: bls12377 hints
