+++ std/algebra/native/fields_bls12377/e2.go
@@ -17,12 +17,9 @@ limitations under the License.
 package fields_bls12377
 
 import (
-	"math/big"
-
 	bls12377 "github.com/consensys/gnark-crypto/ecc/bls12-377"
 	"github.com/consensys/gnark-crypto/ecc/bw6-761/fr"
 
-	"github.com/consensys/gnark/constraint/solver"
 	"github.com/consensys/gnark/frontend"
 )
 
@@ -145,24 +142,6 @@ func (e *E2) Conjugate(api frontend.API, e1 E2) *E2 {
 	return e
 }
 
-var InverseE2Hint = func(_ *big.Int, inputs []*big.Int, res []*big.Int) error {
-	var a, c bls12377.E2
-
-	a.A0.SetBigInt(inputs[0])
-	a.A1.SetBigInt(inputs[1])
-
-	c.Inverse(&a)
-
-	c.A0.BigInt(res[0])
-	c.A1.BigInt(res[1])
-
-	return nil
-}
-
-func init() {
-	solver.RegisterHint(InverseE2Hint)
-}
-
 // Inverse e2 elmts
 func (e *E2) Inverse(api frontend.API, e1 E2) *E2 {
 
@@ -183,26 +162,6 @@ func (e *E2) Inverse(api frontend.API, e1 E2) *E2 {
 	e.assign(res[:2])
 
 	return e
-}
-
-var DivE2Hint = func(_ *big.Int, inputs []*big.Int, res []*big.Int) error {
-	var a, b, c bls12377.E2
-
-	a.A0.SetBigInt(inputs[0])
-	a.A1.SetBigInt(inputs[1])
-	b.A0.SetBigInt(inputs[2])
-	b.A1.SetBigInt(inputs[3])
-
-	c.Inverse(&b).Mul(&c, &a)
-
-	c.A0.BigInt(res[0])
-	c.A1.BigInt(res[1])
-
-	return nil
-}
-
-func init() {
-	solver.RegisterHint(DivE2Hint)
 }
 
 // DivUnchecked e2 elmts
